pyproject.toml
[project]
name = "rent-a-thing-backend"
version = "0.1.0"
description = "Rent-a-Thing — Hyperlocal Rental Marketplace backend (FastAPI, PostgreSQL, Redis, Celery)."
authors = [
    { name = "Rent-a-Thing Team" }
]
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "SQLAlchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0",
    "orjson>=3.10.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "types-redis",
    "types-python-jose",
]

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

